requirements.txt
flask
requests
python-dateutil
Pillow
numpy
python-dotenv
